finallly got models working
