+++ models/login.js
@@ -1,15 +1,14 @@
-
 var Sequelize = require("sequelize");
 var sequelize = require("../config/connection.js");
 
 var login = sequelize.define("login", {
-    loginEmail: {
-      type: Sequelize.STRING
-    },
-    loginPassword: {
-      type: Sequelize.STRING 
-    }
-  }, {
+  loginEmail: {
+    type: Sequelize.STRING
+  },
+  loginPassword: {
+    type: Sequelize.STRING
+  }
+}, {
     timestamps: false
   });
