+++ package.json
@@ -16,10 +16,12 @@
   "bugs": {
     "url": "https://github.com/kjesek/bracketMayhem/issues"
   },
-  "homepage": "https://github.com/kjesek/bracketMayhem#readme",
+  "homepage": "https://github.com/kjesek/bracketMayhem#index",
   "dependencies": {
     "body-parser": "^1.18.2",
     "express": "^4.16.2",
+    "fs": "0.0.1-security",
+    "mysql2": "^1.5.2",
     "path": "^0.12.7",
     "sequelize": "^4.35.2"
   }
