+++ public/index.html
@@ -1,1 +1,7 @@
 
+<!DOCTYPE>
+<html>
+    <head>
+        <title>hey</title>
+    </head>
+</html>
