+++ routes/api-routes.js
@@ -1,30 +1,19 @@
-var liveResult = require("../models/liveResults.js");
-var login = require("../models/logins.js");
-var userData = require("../models/userData.js");
+var liveResult = require("../models/liveResult.js");
+var login = require("../models/login.js");
+var userData = require("../models/userBracket.js");
 
 module.exports = function (app) {
     app.get("/api/liveResults", function (req, res) {
-        liveResult.findAll({
-            where: {
-                waitlisted: false
-            }
-        }).then((results) => res.json(results));
+        liveResult.findAll({}).then((results) => res.json(results));
     });
 
     app.get("/api/logins", function (req, res) {
         login.findAll({
-            where: {
-                waitlisted: false
-            }
         }).then((results) => res.json(results));
     });
 
     app.get("/api/userBrackets", function (req, res) {
-        userBracket.findAll({
-            where: {
-                waitlisted: false
-            }
-        }).then((results) => res.json(results));
+        userBracket.findAll({ }).then((results) => res.json(results));
     });
 
 // Need post route for userBrackets table
