+++ routes/html-routes.js
@@ -4,7 +4,7 @@ var path = require("path");
 
 module.exports = function(app) {
 
-  app.get("/tables", function(req, res) {
+  app.get("/index", function(req, res) {
     res.sendFile(path.join(__dirname, "../public/index.html"));
   });
