+++ schema.sql
@@ -1,0 +1,2 @@
+DROP DATABASE IF EXISTS `jsj3x7apas6k575e`;
+CREATE DATABASE `jsj3x7apas6k575e`;
